Fix A005: rename algorithm/signal.py to convolution.py
The module contained only the Convolution class; the new name matches
its content and avoids shadowing Python's stdlib `signal` module.
Test file and importers updated accordingly.
